Homework16
It was verified that after clicking on the registration button, I was redirected from https://qa.koel.app/ to the https://qa.koel.app/registration page
