+++ src/test/java/Homework16.java
@@ -1,0 +1,34 @@
+import org.openqa.selenium.By;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.WebElement;
+import org.openqa.selenium.chrome.ChromeDriver;
+import org.openqa.selenium.chrome.ChromeOptions;
+import org.testng.Assert;
+import org.testng.annotations.Test;
+
+import java.time.Duration;
+
+public class Homework16 extends BaseTest {
+        @Test
+        public void registrationNavigation() {
+            ChromeOptions options = new ChromeOptions();
+            options.addArguments("--remote-allow-origins=*");
+
+            WebDriver driver = new ChromeDriver(options);
+            driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(10));
+
+            String url = "https://qa.koel.app/";
+            driver.get(url);
+            WebElement registrationButton = driver.findElement(By.cssSelector("a[href='registration']"));
+            registrationButton.click();
+            driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(10));
+            WebElement registrationLink = driver.findElement(By.cssSelector("input[type=email]"));
+            Assert.assertTrue(registrationLink.isDisplayed());
+            driver.quit();
+
+
+
+        }
+
+
+}
