+++ src/test/resources/TestNG.xml
@@ -7,7 +7,7 @@
         <classes>
             <class name="LoginTests"/>
             <class name="BaseTest"/>
-            <class name="HomeWork16"/>
+            <class name="Homework16"/>
         </classes>
     </test>
 </suite>
